% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fastq_urls.R
\name{get_fastq_urls}
\alias{get_fastq_urls}
\title{Retrieve FASTQ download URLs from ENA}
\usage{
get_fastq_urls(srr)
}
\arguments{
\item{srr}{Character vector of SRR accessions.}
}
\value{
A data frame containing SRR accessions and FASTQ URLs.
}
\description{
Given one or more SRA run accessions, retrieve FASTQ download
URLs from the European Nucleotide Archive (ENA).
}
\examples{
\dontrun{
get_fastq_urls("SRR123456")
}

}
